pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "bananaslides"
version = "0.1.0"
description = "Rebuild editable PowerPoint slides from slide images."
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"
dependencies = [
  "numpy>=1.26",
  "onnxruntime>=1.19",
  "opencv-python-headless>=4.9",
  "Pillow>=10.0",
  "pypdfium2>=5.7.0",
  "python-pptx>=0.6.23",
  "rapidocr-onnxruntime>=1.2.3",
  "typer>=0.12",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2",
  "ruff>=0.6",
  "fastapi>=0.115",
  "httpx>=0.27",
  "python-multipart>=0.0.9",
  "uvicorn>=0.30",
]
web = [
  "fastapi>=0.115",
  "httpx>=0.27",
  "python-multipart>=0.0.9",
  "uvicorn>=0.30",
]

[project.scripts]
bananaslides = "bananaslides.cli:app"
bananaslides-web-api = "bananaslides_webapi.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/bananaslides", "src/bananaslides_webapi"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100
